Cherry-pick IPv6 fixes from Heapster
This ensures that we can connect to Kubelets via IPv6 addresses.
It's roughly a cherry-pick of
kubernetes/heapster@a97456be4a936ccc0106f46e0023de5d7f2ce327.

+++ metrics/sources/summary/configs.go
@@ -25,7 +25,7 @@ func GetKubeletConfig(baseKubeConfig *rest.Config, port int, portIsInsecure bool
 	kubeletConfig := &KubeletClientConfig{
 		// TODO: deprecate and remove this option
 		PortIsInsecure: portIsInsecure,
-		Port:           uint(port),
+		Port:           port,
 		RESTConfig:     baseKubeConfig,
 	}
 
@@ -34,7 +34,7 @@ func GetKubeletConfig(baseKubeConfig *rest.Config, port int, portIsInsecure bool
 
 type KubeletClientConfig struct {
 	PortIsInsecure bool
-	Port           uint
+	Port           int
 	RESTConfig     *rest.Config
 }
